fix: create MapContainer component using forward-ref
