+++ pages/map.tsx
@@ -1,20 +1,9 @@
 import React from "react";
-import styled from "@emotion/styled";
 import { 강남역 } from "../lib/constants";
 import useSWR from "swr";
 import { mockClient } from "../lib/api";
 import { IRestaurants } from "../types";
-
-const Styled = {
-  MapContainer: styled.div`
-    height: 100vh;
-
-    .overlay-background {
-      background-color: #fff;
-      padding: 12px;
-    }
-  `,
-};
+import MapContainer from "../components/map/MapContainer";
 
 declare global {
   interface Window {
@@ -24,15 +13,20 @@ declare global {
 
 function Map() {
   const kakaoMap = React.useRef<HTMLDivElement>(null);
+
   const { data: { data: restaurants } = {} } = useSWR<IRestaurants>(
     "/restaurants",
     (url) => mockClient.get(url)
   );
 
   const contentTemplate = (name: string, hashTags: string[]) => `
-    <div class="overlay-background">
-      <div class="overlay-title">${name}</div>
-      <div class="overlay-hashtags">${hashTags.map((hashTag) => hashTag)}</div>
+    <div class="overlay-background" onclick="https://apis.map.kakao.com/web/sample/customOverlay1/" >
+      <a href="https://map.kakao.com/link/map/11394059" target="_blank">
+        <div class="overlay-title">${name}</div>
+        <div class="overlay-hashtags">${hashTags.map(
+          (hashTag) => hashTag
+        )}</div>
+      </a>
     </div>
   `;
 
@@ -60,7 +54,7 @@ function Map() {
     });
   }, [restaurants]);
 
-  return <Styled.MapContainer ref={kakaoMap} />;
+  return <MapContainer ref={kakaoMap} />;
 }
 
 export default Map;
